#16397 move connection setting to the separate connection page; remove constraints comments reading
